Tried to implement mediapipe, we'll see
My first attempt at implementing mediapipe. I made a python module which
contains a class with methods specific to mediapipe, and I use them in
my node handcv.py. I'm not sure if things are going to work and there's
some more work to get everying up and running in the launch file, but all
the pieces are there to get this working in under an hour, hopefully. the
mediapipe functions don't autocomplete in vscode but i'm following the tutorials
precisely.

Added a readme file with a tiny bit of info, as well as a citations.txt where
i reference cv_bridge and mediapipe documentation.

Also added gitignore file.