Added library for realsense and implemented code
Put realsense code from pen project into a library and implemented
it into my ros node. Next I will calculate the centroid of the hand
using the 21 landmarks for the hand.

For some reason mediapipe tries to figure out a z coordinate, which
I think is the location of the hand in the axis in and out of the
camera. This can't be accurate, so I'm going to throw the value away
and rely on depth from the real sense.