MediaPipe is working with realsense camera.
Using the ros wrapper for the realsense camera, I have mediapipe working.
Next I am going to calculate the centroid of the user's hand using
mediapipe's landmarks, and the publish a marker showing the 3D position
of the hand in rviz. I also want to display it on the image.