Fixed an anti patter.
Matt said in class that it's an anti pattern to have to call a function
immediately after creating an object in order to initialize the object.
That is, doing this outside of the class is an antipattern. I realized
that I was doing this so I immediately came in a fxed it.