First iteration of moveit c++ builds, no warnings
I followed the guide on the moveit website to do this, although they
don't use classes just like the official ros documentation doesn't.
Makes me wonder, should I always be using classes just like ROS code in
python, or are C++ patterns different? I think that this will probably
work with the launch files I wrote, but we'll see. After I write
this file, I want to work on getting the reading from the camera more
steady, then I'll be ready to test the naiive approach. Need to accomplish
this by Saturday 1/27 or else I will be in trouble for the harder stuff
down the line.