I think I am reinventing the wheel
I think that not using moveit_cpp is a mistake here, and that I should
transition back to trying to use moveit_cpp everything. Matt seemed to
think that implementing it in one night was doable, and I think that
going through all of the other tutorials that build to moveit_cpp would
take a lot longer than that. Plus, the tutorial for
motion_planning_pipeline is nonexistent, so it leads me to believe maybe
the moveit developers don't think it's super necessary. They should still
have built it out though, no example code???

Seems like moveit_cpp can do everything I'm trying to do in an easier
way.