Launch file works and so does cpp file
Finally, we have something created and launching in rviz. I need to get
the paths to plan still, and I'm just using the moveit tutorial files
for the franka so maybe I'm going to have to do a bunch of extra work
to get it to work on the real franka. Will have to ask Matt this week.