trying to get franka launch stuff to work again
This has been very tough to figure out. it seems like the franka created
launch files and other files are made for the move group node, but I think
that they are the correct and most up to date files to use. on the other
hand, the moveit_cpp tutorials do things completely different. I also
don't know where moveit_cpp ends and the moveit cpp api begins and that
makes it hard to know what objects I should be creating.