Can plan paths and execute them accurately
I now know how to plan and execute a path to a point based on a point in
space and a quaternion. I should write some extra functions for planning
based on specific joint positions, and also a function for planning and
executing sequentially without having to make two service calls.