Added a hand_interface package, and started publishing waypoints
Started publishing waypoints from the handcv node for franka_teleop to
receive and handle. I think that I should make another node to handle
the messages from handcv, and make service calls to franka_teleop for
moving the robot around. I'm not sure how much slower this will be
compared to handling the waypoint messages inside franka teleop, but
considering franka_teleop runs on the robot's computer, it makes sense
to me maybe to handle the messages on my computer, and then control the
robot also with a node on my computer.