I'm about to change some pretty significant stuff
The project is about to not work like it used to work, so they demo I
recorded over the weekend will no longer be possible without giong back
into the code and editing it. However I want to see if this is possible
as faast as possible, using the moveit_servo stuff that is. I'm also
bringing in some of the servo commands into this package so I can
customize them.