Arm is jittery
The arm is jittery and doesn't move very far when i move my hand, but
it looks pretty nice. I need to figure out how to put this on the actual
arm tomorrow, and how to filter out certain jittery stuff. I also should
maybe think about a constant velocity to move the arm at, rather than
letting the raw change in position of the hand determine that.