simulation works great with franka_servo
I've implemented PID control to correct for error in the position of the
end-effector. I define a desired position of the end-effector, and send
my servo node linear pose commands based off of where the end effector
currently is. Right now I'm using a PD control loop in a timer callback
to do this. This accounted for a lot of the jitter I was seeing
yesterday and it looks quite nice.