This might work on real robot? and fixed joint_states.yaml
I'm pretty sure that this *could* work on the real robot. There are some
really silly issues with the joint state broadcaster starting after the
panda_arm_controller, and then the original joint states for the robot
are all zeros and the robot is standing straight up. I think I can fix
this by delaying the start of the franka_servo node from within the node
for a little less than a second. This seems to fix things. This was
causing the robot to jerk around like crazy when I was in the lab, and
also when it starts up in that state i can't send any commands to it or
fix what has happened and I'm not sure why. It might be because it's in
a collision state.