Many to gesture recognition
Removed the left hand from the equation entirely. Now all control is
done with the right hand. Need to think more about if there's a way to
reset the offset of the logic when moving my right hand really fast, but
for now that just did not work very well because mediapipe is not so
when you move fast.

I implemented a publisher that has the positions of the wrist and all of
the tips of the fingers so that I can measure angles. This could be used
later for fine gripper force control.

Implemented thumbs up for start tracking, and thumbs down for stop
tracking.

Implemented pointing up when the gripper is closed to tell the gripper
to start exerting more force, up to a maximum amount of force (10N).

Increase the speed and distance with which the gripper moves when a user
mvoes their hand. Also changed the logic around the hand position noise
threshold.