Added orientation control to the code
Now I can control the orientation of the end effector when grabbing
objects, so no more drifting during the course of operation. I had to
figure out how the servo code accepts rotation increments, then
translate this back all the way in to my cv_franka_bridge file and
figure out how to create euler angle increments for each of the
cartesian axes whenever the quaternion is not what I want it to be.

I also made it so the gripper will (hopefully) stay tangent to the base
link of the robot, instead of always having the quaternion x=1 y=0 z=0
w=0.