More fixes prior to MSI demo
Also, added in something that might help with the jerkiness into
cv_franka_bridge. If the distance is above the upper distance threshold,
if I set the new offset to the waypoint we just received, that could
help? Or not, the robot would still have momentum, and just because for
one iteration I change the offset to something else doesn't mean that
will help the robot slowly stop and change directions. I don't know, but
worth a shot.

I had another idea for fixing the jerkiness, which was that I could have
some kind of momentum term which the change in robot position would have
to slowly chip away from. This would have to be a three vector of some
kind I think. I think this would probably work better on the real robot.