Added a bounding box for end effector
Now, the end effector is not allowed to move outside of a bounding box
that I control. This ensures that the arm will not collide with the
table below it or go anywhere that I don't want it to go. I'm going to
make it so that these bounds can be set with arguments