integrated new franka_ros2 repo with docker and the rest of my ros env
all of my packages, and the franka packages, can see each other and are
being built together. I can launch all of the proper controllers and see
the robot in rviz, with only some warnings in the console.

next steps are to continue improving and connecting the rest of the software,
before thinking about adding new features.