got servo stuff to work in the launch files, working on finishing
still need to tie things together but i was worried about the servo
stuff and it looks like i got it to work. dont know if this will work on
a robot, maybe i can ask one of the msr students to help try and run it
on their stuff.

will need to find a way to switch fr3 to panda easily